feat: Added and styled Navbar
Description: Added a Navbar component  which provides navigation links and applied styles for the Navbar. The Navbar is responsive, and it includes icons for Home, User, Data, and Briefcase. Each link is appropriately styled and positioned within the Navbar. This Navbar is integrated into the project's layout for easy navigation.